Move `save_config` function to a separate module
Refactored the `save_config` function by moving it from `server.py` to `app/utils/settings/save_config.py` for better modularity and organization.
